backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
replicate>=0.22.0
stripe>=7.8.0
boto3>=1.34.0
Pillow>=10.0.0
httpx>=0.26.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
